requirements.txt
asgiref==3.3.1
certifi==2021.5.30
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==2.10
mysqlclient==2.0.3
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.6
